Fix broken image, refactored

+++ deploy.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
+
 set -e # halt script on error
+
 # hugo server --watch &
 #  Install hugo:
 #  hugo version # https://github.com/spf13/hugo/releases
@@ -32,24 +34,36 @@ fi
 
 #echo -e "\nSTART: Spell check contents..."
 #find content -name "*.md" -execdir aspell --mode=html --warn check {} \;
+
 find content -name "*.md.bak" -delete
 find content -name "*~" -delete
-echo 'TODO: grep -R -e "/`grep status archetypes/default.md`/" content/'
+
+# grep --recursive -e png -e jpg -e jpeg -e gif content/ > bugs.txt
+
+_DRAFT=$(      grep --ignore-case --recursive 'draft' content/ | grep --ignore-case 'true')
+_PLACEHOLDER=$(grep --ignore-case --recursive --extended-regexp 'Placeholder' content/)
+
+if [[ ${_PLACEHOLDER} || ${_DRAFT} ]]; then
+  echo "ERROR: found Placeholder or draft:true in content/"
+  echo -e "${_PLACEHOLDER}\n${_DRAFT}\n"
+  exit 8
+fi
 
 echo -e "\nSTART: building project..."
 rm -rf public/*
-hugo --verbose
-if [[ $? -gt 0 ]]; then
-  echo "ERROR: build, exiting."; exit 1;
+
+if ! hugo --verbose; then
+  echo "ERROR: hugo build, exiting."
+  exit 9
 fi
 
 # Test the generated HTML
 # https://github.com/gjtorikian/html-proofer
 if [[ ${TEST} ]]; then
   echo -e "\nSTART: test..."
-  bundle exec htmlproof --verbose ./public
-  if [[ $? -gt 0 ]]; then
-    echo "ERROR: test, exiting."; exit 1;
+  if ! bundle exec htmlproof --verbose ./public; then
+    echo "ERROR: test, exiting."
+    exit                               1
   fi
 else
   echo -e "\n SKIP: test..."
@@ -59,9 +73,9 @@ echo -e "\033[0;32mDeploying updates to GitHub...\033[0m"
 
 # Commit changes.
 cd public && git add --all
-msg="rebuilding site `date`"
-if [ $# -eq 1 ]
-  then msg="$1"
+msg="rebuilding site $(date)"
+if [ $# -eq 1 ]; then
+  msg="$1"
 fi
 git commit -m "$msg"
